requirements.txt
openai 
google-auth 
google-auth-oauthlib 
google-api-python-client 
python-dateutil
loguru